Initial version of Franca product build. Tycho build will produce zip files with Franca IDEs for several platforms.
The primary reason for building these is to control the dependencies of the Franca features. Franca will still be more useful in combination with other tools integrated on some Eclipse IDE platform.

Note: The Mac-OS build cannot be started currently. A „chmod +x“ on the executable „eclipse“ file will be needed (should be part of the product build).