Ensure that „mvn clean“ is deleting everything which should be deleted.
The major step is to centralize the configuration of the maven-clean-plugin into the parent pom.

Special cases: If one plugin generates code into a src-gen folder of another plug-in, the former is responsible for cleaning. I.e., the „receiving“ plug-ins maven-clean-plugin configuration has to be cut down in order to avoid cleaning files in the src-gen folder after they have been generated during a regular build.