Switched implementation of Franca annotations, colon character was blocked by lexer before.
Now, the colon character can be used by other parser rules (and will be actually).
There are some drawbacks for the new implementation of annotations. The most important
drawback is that content assist for the annotation tag isn’t available anymore. But this might
be reactivated by a more sophisticated implementation lateron.