Allow arithmetic expressions as enumerator value (first part of issue 52).
The expressions are type-checked. We allow everything which
evaluates to an integer. We also allow string values, but this is
deprecated and will lead to a warning.

There is no evaluation of the value expressions, yet.
This has to be implemented in order to make the Franca2DBus
transformation work correctly.

There are also no hex numbers supported yet.