Switched fdepl from an inferred to an explicitly imported Ecore model.
This was necessary because the Xtext Ecore model inferrer treats artifical parser rules as real ones if the rule is called from another artificial parser rules.
This is needed to introduce a subclass for FDElement, but without generating a parser rule for the subclass (because it never could be parsed correctly).