Added deployment model validation for conflicting property declarations (motivated by issue 101).
The conflicts result exclusively from property declarations with the same name. Among each group of declarations with equal name, several check rules are applied. These rules are due to  restrictions in the generated Java code for PropertyAccessors. The goal of the new validation is to avoid any problems in the generated PropertyAccessor API.