Support OMG IDL multiple interface inheritance with a workaround as long as Franca IDL does not support it.
In the OMG IDL transformation, we map the first base interface to the Franca IDL base interface. All other base interface (in case of multiple inheritance) are flattened, i.e., their attributes and methods are cloned into the derived Franca interface. This is done transitively, i.e., all elements the flattened base interface inherits from its base will be flattened, too. We added are unit tests for this.

A structured comment is added to the Franca interface elements in order to document the flattening of the resulting attributes and members.

This is part of issue #183.