Proper handling of OMG IDL object references (i.e., "Object") type in transformation (issue #195).
The OMG IDL "Object" datatype is mapped to a Franca IDL typedef named "Object" which is an alias for the "undefined" type. The implementation of the object reference is depending on the platform implementation. Usually, a Franca interface is the type of a port element, and a providing component could have multiple ports. In CORBA, it seems that there is no distinction between an "object" and an interface instance. Thus, an object reference in CORBA (and therefore in OMG IDL) seems to map the concept of "port" in Franca-based systems. Currently, in OMGIDLBase.fidl the Object type is an alias of "undefined". If the specific platform uses some specific datatype for referencing objects (or port instances), the type can be adapted in OMGIDLBase.fidl and the adapted base types file has to be used. E:g., if port instances are referenced via strings, OMGIDLBase.fidl should contain "typedef Object as string".