Moved p2 repositories for 0.8.x releases to new releases_old folder.
During installation using the "releases" p2 repo, features from the old Franca version 0.8.11 are visible, although "Show only latest version" is checked in the IDE. This is because the feature structure has been redesigned from 0.8.11 to 0.9.0. Users installing Franca will become mixed up by this and the installation will not work when erroneously selecting 0.8.x versions and newer versions.

The 0.8.x releases have not been deleted, but moved to a "releases_old" folder to archive them.