Removed Franca 0.9.2 from p2 repository.
Franca 0.9.2 is no drop-in replacement for 0.9.1, thus, it is not backward compatible. The IDL is actually compatible, but it won’t work with CommonAPI C++. In order to allow providing a compatible 0.9.x release (which _is_ compatible with CommonAPI C++) with back-ported fixes and features from later versions, we are withdrawing the 0.9.2 release.