Fix formatting and wrong inserting of textual model changes at the beginning of the file (issue #219).
This is done by forcing a change on the FDRootElement which contains the actual model modifications.

The „insertion at location 0“ issue was caused if the FDPropertySet is the top model element that has been modified by the quickfix. Xtext’s semantic modification computed a wrong insertion region for there. Using the enclosing FDElement as modified object did help for this.

The formatting issue could be solved by going up the AST even more and using the FDRootElement. Then the indentation after inserting the quickfix modification usually is correct.